the code was not identical to the one in my .Rmd file because I made a type (2 instead of 1).
I have to revise the entire Curate module, because from the time of writing it to now I have much progress in my understanding of Shiny. Also, There should not be manually typed indexed in there but only things that directly refer to user-inputs. Like this, it's the user that makes the mistake and not my code.